requirements.txt
streamlit==1.31.1
scikit-learn==1.4.2
joblib==1.3.2
numpy==1.26.4
Pillow==9.4.0
streamlit-option-menu==0.3.12 
tensorflow==2.16.1
keras==3.1.1
zipp==3.17.0